+++ tests/Core/MainTest.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace MiBo\Prices\Tests;
 
+use DateTime;
 use MiBo\Prices\Calculators\PriceCalc;
 use MiBo\Prices\Price;
+use MiBo\Prices\Taxonomies\AnyTaxonomy;
 use MiBo\Prices\Units\Price\Currency;
 use MiBo\Properties\Calculators\UnitConvertor;
 use MiBo\VAT\Enums\VATRate;
-use MiBo\VAT\Resolvers\ProxyResolver;
+use MiBo\VAT\Manager;
 use MiBo\VAT\VAT;
 use PHPUnit\Framework\TestCase;
 
@@ -51,11 +53,11 @@ class MainTest extends TestCase
     public function test(): void
     {
         $prices = [
-            new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE")),
-            new Price(15, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9704 00 00", "CZE")),
-            new Price(20, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(25, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(30, Currency::get("CZK"), ProxyResolver::retrieveByCategory("07", "CZE")),
+            new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE")),
+            new Price(15, Currency::get("CZK"), $this->retrieveVATByCategory("9704 00 00", "CZE")),
+            new Price(20, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(25, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(30, Currency::get("CZK"), $this->retrieveVATByCategory("07", "CZE")),
         ];
 
         $price = clone $prices[0];
@@ -70,7 +72,7 @@ class MainTest extends TestCase
         $this->assertSame(119.75, $price->getValueWithVAT());
 
         $price->add(
-            new Price(1, Currency::get("EUR"), ProxyResolver::retrieveByCategory("07", "SVK"))
+            new Price(1, Currency::get("EUR"), $this->retrieveVATByCategory("07", "SVK"))
         );
 
         $this->assertSame(135, $price->getValue());
@@ -96,7 +98,7 @@ class MainTest extends TestCase
     {
         PriceCalc::setValueOfVATClosure(null);
         PriceCalc::getValueOfVAT(
-            new Price(0, Currency::get('CZK'), ProxyResolver::retrieveByCategory('06', 'CZE'))
+            new Price(0, Currency::get('CZK'), $this->retrieveVATByCategory('06', 'CZE'))
         );
 
         PriceCalc::setValueOfVATClosure(function(): int {
@@ -106,7 +108,7 @@ class MainTest extends TestCase
         $this->assertEquals(
             0,
             PriceCalc::getValueOfVAT(
-                new Price(0, Currency::get('CZK'), ProxyResolver::retrieveByCategory('06', 'CZE'))
+                new Price(0, Currency::get('CZK'), $this->retrieveVATByCategory('06', 'CZE'))
             )
         );
     }
@@ -120,13 +122,13 @@ class MainTest extends TestCase
      */
     public function testAddingCombined(): void
     {
-        $price = new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
-        $price->add(new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("07", "CZE")));
+        $price = new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
+        $price->add(new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("07", "CZE")));
 
         $this->assertSame(20, $price->getValue());
         $this->assertSame(21.5, $price->getValueWithVAT());
 
-        $price2 = new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
+        $price2 = new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
         $price2->add($price);
 
         $this->assertSame(30, $price2->getValue());
@@ -143,11 +145,11 @@ class MainTest extends TestCase
     public function testAnyRateSubtraction(): void
     {
         $prices = [
-            new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE")),
-            new Price(15, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9704 00 00", "CZE")),
-            new Price(20, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(25, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(30, Currency::get("CZK"), ProxyResolver::retrieveByCategory("07", "CZE")),
+            new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE")),
+            new Price(15, Currency::get("CZK"), $this->retrieveVATByCategory("9704 00 00", "CZE")),
+            new Price(20, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(25, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(30, Currency::get("CZK"), $this->retrieveVATByCategory("07", "CZE")),
         ];
 
         $price = clone $prices[0];
@@ -156,7 +158,11 @@ class MainTest extends TestCase
             $price->add($p);
         }
 
-        $discount = new Price($price->getValue(), Currency::get("CZK"), VAT::get("CZE", VATRate::ANY));
+        $discount = new Price(
+            $price->getValue(),
+            Currency::get("CZK"),
+            VAT::get('CZE', VATRate::ANY, AnyTaxonomy::get(), new DateTime())
+        );
 
         $price->subtract($discount);
 
@@ -172,24 +178,24 @@ class MainTest extends TestCase
      */
     public function testSubtraction(): void
     {
-        $price = new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
-        $price->subtract(new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE")));
+        $price = new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
+        $price->subtract(new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE")));
 
         $this->assertSame(0, $price->getValue());
         $this->assertSame(0, $price->getValueWithVAT());
 
-        $price2 = new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
+        $price2 = new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
         $price2->subtract($price);
 
         $this->assertSame(10, $price2->getValue());
         $this->assertSame(11.5, $price2->getValueWithVAT());
 
         $prices = [
-            new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE")),
-            new Price(15, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9704 00 00", "CZE")),
-            new Price(20, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(25, Currency::get("CZK"), ProxyResolver::retrieveByCategory("2201", "CZE")),
-            new Price(30, Currency::get("CZK"), ProxyResolver::retrieveByCategory("07", "CZE")),
+            new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE")),
+            new Price(15, Currency::get("CZK"), $this->retrieveVATByCategory("9704 00 00", "CZE")),
+            new Price(20, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(25, Currency::get("CZK"), $this->retrieveVATByCategory("2201", "CZE")),
+            new Price(30, Currency::get("CZK"), $this->retrieveVATByCategory("07", "CZE")),
         ];
 
         $price = clone $prices[0];
@@ -228,8 +234,8 @@ class MainTest extends TestCase
      */
     public function testAddingForeignPrice(): void
     {
-        $priceCZK = new Price(10, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
-        $priceEUR = new Price(10, Currency::get("EUR"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
+        $priceCZK = new Price(10, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
+        $priceEUR = new Price(10, Currency::get("EUR"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
 
         [
             $price,
@@ -246,7 +252,9 @@ class MainTest extends TestCase
     {
         parent::setUp();
 
-        ProxyResolver::setResolver(VATResolver::class);
+        $vatHelper = new VATResolver();
+
+        PriceCalc::setVATManager(new Manager($vatHelper, $vatHelper, $vatHelper));
 
         // Setting conversion rate between CZK and EUR => 1 EUR = 25 CZK
         UnitConvertor::$unitConvertors[\MiBo\Prices\Quantities\Price::class] = function(Price $price, Currency $unit) {
@@ -267,5 +275,10 @@ class MainTest extends TestCase
     {
         PriceCalc::setValueOfVATClosure(null);
         parent::tearDown();
+    }
+
+    protected function retrieveVATByCategory(string $category, string $country): VAT
+    {
+        return PriceCalc::getVATManager()->retrieveVAT(new TestingClassification($category), $country);
     }
 }