+++ tests/Core/PriceMultiplicationTest.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace MiBo\Prices\Tests;
 
+use MiBo\Prices\Calculators\PriceCalc;
 use MiBo\Prices\Price;
 use MiBo\Prices\Units\Price\Currency;
 use MiBo\Properties\Calculators\UnitConvertor;
-use MiBo\VAT\Resolvers\ProxyResolver;
+use MiBo\VAT\Manager;
+use MiBo\VAT\VAT;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -36,7 +38,7 @@ class PriceMultiplicationTest extends TestCase
      */
     public function test(): void
     {
-        $price = new Price(100, Currency::get("CZK"), ProxyResolver::retrieveByCategory("9705 00 00", "CZE"));
+        $price = new Price(100, Currency::get("CZK"), $this->retrieveVATByCategory("9705 00 00", "CZE"));
         $price->multiply(2);
 
         $this->assertSame(200, $price->getValue());
@@ -52,7 +54,7 @@ class PriceMultiplicationTest extends TestCase
         $this->assertSame(200, $price->getValue());
         $this->assertSame(30.0, $price->getValueOfVAT());
 
-        $price->add(new Price(100, Currency::get("CZK"), ProxyResolver::retrieveByCategory("07", "CZE")));
+        $price->add(new Price(100, Currency::get("CZK"), $this->retrieveVATByCategory("07", "CZE")));
 
         $this->assertSame(300, $price->getValue());
         $this->assertSame(30.0, $price->getValueOfVAT());
@@ -70,7 +72,9 @@ class PriceMultiplicationTest extends TestCase
     {
         parent::setUp();
 
-        ProxyResolver::setResolver(VATResolver::class);
+        $vatHelper = new VATResolver();
+
+        PriceCalc::setVATManager(new Manager($vatHelper, $vatHelper, $vatHelper));
 
         // Setting conversion rate between CZK and EUR => 1 EUR = 25 CZK
         UnitConvertor::$unitConvertors[\MiBo\Prices\Quantities\Price::class] = function(Price $price, Currency $unit) {
@@ -82,5 +86,10 @@ class PriceMultiplicationTest extends TestCase
 
             return $price->getNumericalValue()->divide(25);
         };
+    }
+
+    protected function retrieveVATByCategory(string $category, string $country): VAT
+    {
+        return PriceCalc::getVATManager()->retrieveVAT(new TestingClassification($category), $country);
     }
 }